# docs/openapi.json
{"openapi":"3.0.3","info":{"title":"TRUSTMESH API","version":"1.0.0","description":"Local-first prototype API. Device endpoints take HMAC envelopes (see IOT_INTEGRATION.md)."},"components":{"schemas":{}},"paths":{"/api/v1/auth/personas":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/api/v1/auth/dev-login":{"post":{"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["persona"],"properties":{"persona":{"type":"string"}}}}}},"responses":{"200":{"description":"Default Response"}}}},"/api/v1/auth/nonce":{"post":{"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["address"],"properties":{"address":{"type":"string","pattern":"^0x[0-9a-fA-F]{40}$"}}}}}},"responses":{"200":{"description":"Default Response"}}}},"/api/v1/auth/wallet":{"post":{"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["message","signature"],"properties":{"message":{"type":"string","maxLength":2000},"signature":{"type":"string"}}}}}},"responses":{"200":{"description":"Default Response"}}}},"/api/v1/auth/me":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/api/v1/auth/logout":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/api/v1/demo/state":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/api/v1/demo/{step}":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"step","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/sim":{"get":{"responses":{"200":{"description":"Default Response"}}},"post":{"responses":{"200":{"description":"Default Response"}}}},"/health":{"get":{"tags":["health"],"responses":{"200":{"description":"Default Response"}}}},"/ready":{"get":{"tags":["health"],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/device/hello":{"post":{"tags":["device"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["deviceId","keyVersion","payloadB64","macHex"],"additionalProperties":false,"properties":{"deviceId":{"type":"string","maxLength":64},"keyVersion":{"type":"integer","minimum":1},"payloadB64":{"type":"string","maxLength":6000},"macHex":{"type":"string","maxLength":64}}}}}},"responses":{"200":{"description":"Default Response"}}}},"/api/v1/device/telemetry":{"post":{"tags":["device"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["deviceId","keyVersion","payloadB64","macHex"],"additionalProperties":false,"properties":{"deviceId":{"type":"string","maxLength":64},"keyVersion":{"type":"integer","minimum":1},"payloadB64":{"type":"string","maxLength":6000},"macHex":{"type":"string","maxLength":64}}}}}},"responses":{"200":{"description":"Default Response"}}}},"/api/v1/device/poll":{"post":{"tags":["device"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["deviceId","keyVersion","payloadB64","macHex"],"additionalProperties":false,"properties":{"deviceId":{"type":"string","maxLength":64},"keyVersion":{"type":"integer","minimum":1},"payloadB64":{"type":"string","maxLength":6000},"macHex":{"type":"string","maxLength":64}}}}}},"responses":{"200":{"description":"Default Response"}}}},"/api/v1/device/ack":{"post":{"tags":["device"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["deviceId","keyVersion","payloadB64","macHex"],"additionalProperties":false,"properties":{"deviceId":{"type":"string","maxLength":64},"keyVersion":{"type":"integer","minimum":1},"payloadB64":{"type":"string","maxLength":6000},"macHex":{"type":"string","maxLength":64}}}}}},"responses":{"200":{"description":"Default Response"}}}},"/api/v1/stream":{"get":{"tags":["live"],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/chain/manifest":{"get":{"tags":["chain"],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/chain/status":{"get":{"tags":["chain"],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/chain/txs":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/api/v1/chain/events":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/api/v1/chain/tx/{hash}":{"get":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"hash","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/chain/reindex":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/api/v1/chain/track":{"post":{"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["hash"],"properties":{"hash":{"type":"string","pattern":"^0x[0-9a-fA-F]{64}$"}}}}}},"responses":{"200":{"description":"Default Response"}}}},"/api/v1/overview":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/api/v1/assets":{"get":{"responses":{"200":{"description":"Default Response"}}},"post":{"tags":["assets"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["id","name","location","owner","requireReal"],"properties":{"id":{"type":"string"},"name":{"type":"string","minLength":2,"maxLength":80},"model":{"type":"string","maxLength":80},"description":{"type":"string","maxLength":500},"location":{"type":"string","maxLength":80},"owner":{"type":"string","pattern":"^0x[0-9a-fA-F]{40}$"},"custodian":{"type":"string","pattern":"^0x[0-9a-fA-F]{40}$"},"requireReal":{"type":"boolean"},"requiredFlags":{"type":"integer","minimum":0},"maxEvidenceAge":{"type":"integer","minimum":10,"maximum":86400}}}}}},"responses":{"200":{"description":"Default Response"}}}},"/api/v1/assets/{id}":{"get":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/assets/{id}/enroll-tag":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/assets/{id}/sensor-policy":{"put":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/public/assets/{id}":{"get":{"tags":["public"],"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/public/kit":{"get":{"tags":["public"],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/transfers":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/api/v1/transfers/{id}/challenge":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/actions/{name}":{"post":{"tags":["actions"],"parameters":[{"schema":{"type":"string"},"in":"path","name":"name","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/incidents/{id}/reports":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/incidents/{id}/maintenance-evidence":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/incidents":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/api/v1/devices":{"get":{"responses":{"200":{"description":"Default Response"}}},"post":{"tags":["devices"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["id","provenance","profile"],"properties":{"id":{"type":"string"},"label":{"type":"string","maxLength":80},"provenance":{"enum":["REAL","SIMULATED"]},"profile":{"type":"string","maxLength":32},"caps":{"type":"integer","minimum":0},"assetId":{"type":"string"}}}}}},"responses":{"200":{"description":"Default Response"}}}},"/api/v1/devices/{id}/telemetry":{"get":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/devices/{id}/bind":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/devices/{id}/revoke":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/devices/{id}/rotate-key":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/commands":{"get":{"responses":{"200":{"description":"Default Response"}}},"post":{"tags":["commands"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["deviceId","action"],"properties":{"deviceId":{"type":"string"},"action":{"type":"string"},"params":{"type":"object"}}}}}},"responses":{"200":{"description":"Default Response"}}}},"/api/v1/credentials":{"get":{"responses":{"200":{"description":"Default Response"}}},"post":{"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["holder","role","days"],"properties":{"holder":{"type":"string","pattern":"^0x[0-9a-fA-F]{40}$"},"role":{"enum":["TECHNICIAN","INSPECTOR"]},"days":{"type":"number","exclusiveMinimum":0,"maximum":3650}}}}}},"responses":{"200":{"description":"Default Response"}}}},"/api/v1/credentials/revoke":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/api/v1/organizations":{"get":{"responses":{"200":{"description":"Default Response"}}},"post":{"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["name","kind"],"properties":{"name":{"type":"string","minLength":2,"maxLength":80},"kind":{"type":"string","maxLength":32}}}}}},"responses":{"200":{"description":"Default Response"}}}},"/api/v1/organizations/members":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/api/v1/evidence":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/api/v1/integrity/{eventId}":{"get":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"eventId","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/integrity/sandbox":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/api/v1/integrity/sandbox/{id}/modify":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/integrity/sandbox/{id}/restore":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/api/v1/audit":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/api/v1/diagnostics":{"get":{"responses":{"200":{"description":"Default Response"}}}}}}